Move the request parser BookingInfoValidationRequestBodyParser
